Update naming scheme
