+++ metadata_backend/api/handlers/folder.py
@@ -14,7 +14,7 @@ from ...conf.conf import publisher
 from ...helpers.doi import DOIHandler
 from ...helpers.logger import LOG
 from ...helpers.validator import JSONValidator
-from .api_handlers import RESTAPIHandler
+from .restapi import RESTAPIHandler
 from ..middlewares import get_session
 from ..operators import FolderOperator, Operator, UserOperator
